* p4a-validation:
  accept changes due to terapix bw change (again)
  accept changes due to updated memory bandwidth to fill-in the terapix
  Add a new example to the intraprocedural chapter
  Not a clean checkpoint, but it is Friday night
  accept new more generic code
  accept new handling of reductions
  C_syntax: pml03 added
  accept identation fix
  accept new version including one test with tiling enabled
  accept added comment.

+++ packages/PIPS/validation/Pointers/malloc02.c
@@ -8,7 +8,7 @@ int main(int argc, char *argv[])
   float * px;
 
   pi = (int *) malloc(sizeof(int));
-  px = (float *) malloc(sizeof(int));
+  px = (float *) malloc(sizeof(float));
 
   return 0;
 }
